Update pip to use patched txtorcon version.
The published pip version will crash if we kill it's underlying
Tor process or it will leave Tor processes running if the launch
timeout is reached. That causes problems with launching new Tor
instances in future, so the patched version hacks around this until
meejah let's me know if there's a better alternative.